Move controller code into `AccountLogin` widget
